+++ pixora/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """pixora modules."""
 from .params import PIXORA_VERSION
-from .algorithms import NearestNeighbor, Lanczos, Bilinear
+from .algorithms import NearestNeighbor, Lanczos, Bilinear, Bicubic
 from .converter import Converter, pixelize
 from .errors import PixoraError, PixoraImageError, PixoraValidationError
 
@@ -15,5 +15,6 @@ __all__ = [
     "PixoraValidationError",
     "NearestNeighbor",
     "Lanczos",
-    "Bilinear"
+    "Bilinear",
+    "Bicubic"
 ]