+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `Lanczos` algorithm
 - `Bilinear` algorithm
+- `Bicubic` algorithm
 ### Changed
 - Test system modified
 ## [0.1] - 2026-07-14